requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==4.4.0
    # via
    #   httpx
    #   openai
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
click==8.0.3  # Specify a compatible version of click
    # via
    #   flask
    #   typer
distro==1.9.0
    # via openai
flask==2.0.3
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
openai==1.34.0
    # via -r requirements.in
pydantic==1.9.0
    # via
    #   -r requirements.in
    #   openai
python-dotenv==0.20.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.4
    # via openai
typer==0.4.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
werkzeug==2.0.1
    # via flask